# packages/contracts/openapi/novel-factory.v1.yaml
openapi: 3.1.0
info:
  title: Novel Factory API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /books/{bookId}:
    get:
      operationId: getBook
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /books/{bookId}/chapters:
    get:
      operationId: listBookChapters
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book chapters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
...
  /extraction-runs/{runId}:
    get:
      operationId: getExtractionRun
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extraction run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Extraction run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /extraction-runs/{runId}/report:
    get:
      operationId: getExtractionRunReport
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Extraction run report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
...
  /knowledge-objects/{objectId}/review-actions:
    post:
      operationId: reviewKnowledgeObject
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewActionCommand'
      responses:
        '200':
          description: Review action applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
...
  /novel-projects/{projectId}:
    get:
      operationId: getNovelProject
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Novel project detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Novel project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /workspaces/{workspaceId}/home:
    get:
      operationId: getWorkspaceHome
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace home aggregate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /chapter-plans/{chapterPlanId}:
    get:
      operationId: getChapterPlan
      parameters:
        - in: path
          name: chapterPlanId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chapter plan detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Chapter plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /chapter-plans/{chapterPlanId}/section-plans:
    get:
      operationId: listSectionPlans
      parameters:
        - in: path
          name: chapterPlanId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Section plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
    post:
      operationId: createSectionPlans
      parameters:
        - in: path
          name: chapterPlanId
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Section plans accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
...
  /model-profiles/{modelProfileId}/enable:
    post:
      operationId: enableModelProfile
      parameters:
        - in: path
          name: modelProfileId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model profile enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Model profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /model-profiles/{modelProfileId}/disable:
    post:
      operationId: disableModelProfile
      parameters:
        - in: path
          name: modelProfileId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model profile disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
        '404':
          description: Model profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
components:
  schemas:
...
    ApiSuccessEnvelope:
      $ref: ../schemas/api-success-envelope.schema.json
    ApiErrorEnvelope:
      $ref: ../schemas/api-error-envelope.schema.json
...
    ReviewActionCommand:
      $ref: ../schemas/review-action-command.schema.json
...